Integrated fuel mechanics into the rest of game
-Still trying to make trash collection happen reliably. May need to rework trash collector to not use physics. If the boat is moving too fast, it all just flies out the back.

+++ Assets/Bambi/scriptDashBoard.cs
@@ -28,7 +28,7 @@ public class scriptDashBoard : MonoBehaviour
 		txtTrashAmount.text = $"Trash: {trashCount}";
 	}
 
-	public void updateFuelAmount(int fuelCount)
+	public void updateFuelAmount(float fuelCount)
 	{
 		txtTrashAmount.text = $"Fuel: {fuelCount}";
 	}
